// src/utils/vulnType.d.ts
export interface Result {
  resultsPerPage: number;
  startIndex: number;
  totalResults: number;
  format: string;
  version: string;
  timestamp: Date;
  vulnerabilities: Vulnerability[];
}

export interface Vulnerability {
  cve: Cve;
}

export interface Cve {
  id: string;
  sourceIdentifier: SourceIdentifierEnum;
  published: Date;
  lastModified: Date;
  vulnStatus: VulnStatus;
  descriptions: Description[];
  metrics: Metrics;
  weaknesses: Weakness[];
  configurations: Configuration[];
  references: Reference[];
  evaluatorComment?: string;
  cisaExploitAdd?: Date;
  cisaActionDue?: Date;
  cisaRequiredAction?: string;
  cisaVulnerabilityName?: string;
}